refactor(assembly): update consumer and producer
